requirements.txt
# MovieFetchBot Dependencies
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit>=1.28.0

# HTTP Client
httpx==0.25.2

# Data Validation
pydantic==2.5.2
pydantic-settings==2.1.0

# Data Processing
pandas==2.1.3

# Authentication
python-jose[cryptography]==3.3.0

# File Handling
python-multipart==0.0.6
aiofiles==23.2.1

# Rate Limiting
slowapi==0.1.9

# Session Persistence
extra-streamlit-components>=0.1.60

